# openapi.yaml
openapi: 3.1.0
info:
  title: AgentMail API
  version: 1.0.0
  description: >
    API for giving AI agents their own email inboxes — send, receive, and
    manage threaded conversations programmatically.
  x-organization: agentmail-demo
servers:
  - url: https://agentmail-demo.vercel.app
    description: Production
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: API key (sk_live_...)
  schemas:
...
    Message:
      type: object
      required: [id, thread_id, inbox_email_address, direction, message_id_header, from, to, cc, bcc, headers, status, created_at]
      properties:
        id:
          type: string
          description: RFC 5322 Message-ID (normalized, without angle brackets)
          example: 1234567890.abcdef.12345@yourdomain.com
        thread_id:
          type: string
          example: thread_abc123
        inbox_email_address:
          type: string
          format: email
          example: alice@yourdomain.com
        direction:
          type: string
          enum: [inbound, outbound]
        message_id_header:
          type: string
        in_reply_to:
          type: [string, "null"]
        from:
          type: string
          format: email
        to:
          type: array
          items:
            type: string
            format: email
        cc:
          type: array
          items:
            type: string
            format: email
        bcc:
          type: array
          items:
            type: string
            format: email
        subject:
          type: [string, "null"]
        body_text:
          type: [string, "null"]
        body_html:
          type: [string, "null"]
        headers:
          type: object
          additionalProperties: true
        status:
          type: string
          enum: [sent, received, delivered, bounced]
        created_at:
          type: string
          format: date-time
...
    MessageList:
      type: object
      required: [object, data, has_more]
      properties:
        object:
          type: string
          enum: [list]
        data:
          type: array
          items:
            $ref: "#/components/schemas/Message"
        has_more:
          type: boolean
        next_cursor:
          type: [string, "null"]
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
paths:
...
  /v1/messages:
    post:
      operationId: sendMessage
      summary: Send message
      description: >
        Send an outbound email. Creates a new thread unless in_reply_to_thread_id
        is provided, in which case the message is appended to that thread.
      tags: [Messages]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [inbox_email_address, to]
              properties:
                inbox_email_address:
                  type: string
                  format: email
                  description: Sending inbox.
                to:
                  oneOf:
                    - type: string
                      format: email
                    - type: array
                      items:
                        type: string
                        format: email
                cc:
                  oneOf:
                    - type: string
                      format: email
                    - type: array
                      items:
                        type: string
                        format: email
                bcc:
                  oneOf:
                    - type: string
                      format: email
                    - type: array
                      items:
                        type: string
                        format: email
                subject:
                  type: string
                text:
                  type: string
                  description: Plain-text body. At least one of text or html is required.
                html:
                  type: string
                  description: HTML body. At least one of text or html is required.
                in_reply_to_thread_id:
                  type: string
                  description: Reply in this thread instead of starting a new one.
                headers:
                  type: object
                  additionalProperties: true
                  description: Custom RFC headers to inject.
      responses:
        "201":
          description: Message sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Inbox or thread not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    get:
      operationId: listMessages
      summary: List messages
      description: List messages with optional filters. Returns newest first.
      tags: [Messages]
      parameters:
        - name: inbox_email_address
          in: query
          schema:
            type: string
            format: email
        - name: thread_id
          in: query
          schema:
            type: string
        - name: direction
          in: query
          schema:
            type: string
            enum: [inbound, outbound]
        - name: from
          in: query
          schema:
            type: string
            format: email
        - name: to
          in: query
          schema:
            type: string
            format: email
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: starting_after
          in: query
          description: Cursor for pagination (message ID of the last item).
          schema:
            type: string
      responses:
        "200":
          description: List of messages
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageList"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Inbox not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /v1/messages/{id}:
    get:
      operationId: getMessage
      summary: Get message
      tags: [Messages]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Message object
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /v1/messages/{id}/raw:
    get:
      operationId: getRawMessage
      summary: Get raw message
      description: Returns the raw RFC 822 (.eml) source of a message.
      tags: [Messages]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Raw RFC 822 message
          content:
            text/plain:
              schema:
                type: string
        "404":
          description: Not found or raw source not preserved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...